moved intelligence reducfer to container and implimented direct testing, minor other fixes
